R7, R8

+++ Code/R8_FilterTracks.py
@@ -68,7 +68,7 @@ print(UF.TimeStamp(), bcolors.OKGREEN+"Modules Have been imported successfully..
 print(UF.TimeStamp(),'Loading preselected data from ',bcolors.OKBLUE+input_file_location+bcolors.ENDC)
 data=pd.read_csv(input_file_location,header=0)
 print(UF.TimeStamp(),'Analysing data... ',bcolors.ENDC)
-NoJobs=np.ceil(len(data)/MaxFitTracksPerJob)
+NoJobs=int(np.ceil(len(data)/MaxFitTracksPerJob))
 if Mode=='R':
    print(UF.TimeStamp(),bcolors.WARNING+'Warning! You are running the script with the "Mode R" option which means that you want to vertex the seeds from the scratch'+bcolors.ENDC)
    print(UF.TimeStamp(),bcolors.WARNING+'This option will erase all the previous Seed vertexing jobs/results'+bcolors.ENDC)
